+++ icloudpd/email_notifications.py
@@ -8,11 +8,11 @@ from icloudpd.logger import setup_logger
 
 
 def send_2sa_notification(
-        smtp_email, smtp_password, smtp_host, smtp_port, smtp_no_tls, to_addr
+        smtp_email, smtp_password, smtp_host, smtp_port, smtp_no_tls, to_addr, from_addr = None
 ):
     """Send an email notification when 2SA is expired"""
     to_addr = to_addr if to_addr else smtp_email
-    from_addr = smtp_email if smtp_email else to_addr
+    from_addr = from_addr or (f"iCloud Photos Downloader <{smtp_email}>" if smtp_email else to_addr)
     logger = setup_logger()
     logger.info("Sending 'two-step expired' notification via email...")
     smtp = smtplib.SMTP(smtp_host, smtp_port)
@@ -34,7 +34,7 @@ def send_2sa_notification(
 Two-step authentication has expired for the icloud_photos_downloader script.
 Please log in to your server and run the script manually to update two-step authentication."""
 
-    msg = f"From: iCloud Photos Downloader <{from_addr}>\n" + \
+    msg = f"From: {from_addr}\n" + \
         f"To: {to_addr}\nSubject: {subj}\nDate: {date}\n\n{message_text}"
 
     smtp.sendmail(from_addr, to_addr, msg)